Create rviz2 launch file
Single out rviz2 from the warehouse_bot launch.
This could be useful if we are running robot model package on real
robot. And we want to use rviz to monitor the sensor information.

The reason why rviz launch is not in the simulation package is because
rviz is not very useful if the robot hasn't brought up yet.